A better test for the CoefficientAccessor.
The previous one ignored the fact that a JCOEF is a short int, and as
such had some overflow problems. This one is fine, and the
CoefficientAccessor actually passes it.

As a note to self, this test makes mixed use of malloc and new. Very
bad! I'll have to go back and fix that soon.