A crappy test for doCrop and some structure to test JPEGImage.
Unfortunately, doCrop is lossy by design (the algorithms in use by
stegosaurus positively require the crop to be lossy), which means that
we can't do any formal verification that the crop has worked beyond
ensuring that it doesn't crash.

In the future, the test will probably be expanded to check some basic
info from the cropped image and ensure that it's coherent. I may even
go the full length of cropping an image in another program so that the
test can have something to compare the results of doCrop to.