Yet another Individual implementation to simplify testing a bit.
Yes, welcome to test-a-palooza. This class has no conceivable use other
than its instances being diverse enough to allow us to say in a real
sense that one is fitter than the other (unlike the other one, which
made no use of the Chromosome whatsoever, and so was useless to test
most genetic principles). And because we're in such a test-frenzy, test
cases for a class that will only be used in test cases have also been
included.