Implement a general framework for genetic algorithms.
Basically, this should allow us to just pass in any crossover, mutation
and selection operators, and have a corresponding genetic algorithm
without having to implement the common, abstract, workflow (construct
chromosomes, call simulate, acquire fitness, and so on).

There's also a test for it, which contains an extremely simple case of
running the algorithm for a single generation, as a consistency check,
and a somewhat more complex case of having a genetic algorithm try to
approximante a given constant.