Each JPEGImage now has its own jpeg_decompress_struct.
This is a pretty significant change, as previously we were sharing them
around. While the previous way of doing things was evidently not
thread-safe, it turns out (unsurprisingly) that it is incredibly
inflexible even in a single-threaded context.
This has an associated memory cost, but is ultimately the only way to
go.
Note that jpeg_compress_structs have also been refactored in this
fashion, which may not be strictly necessary and is open to further
change.
Finally, this leaves the jpeg_context as a pretty vestigal structure,
but since it's already there, and we may want to introduce different
construction schemes, it's still there.