An ugly mechanism allowing for a JPEGImage to be re-used.
JPEGImages were never designed as throwaway objects. They just turned
out that way because I suffer from chronic idiocy syndrome. So that had
to be fixed. I'm still not certain I've got it down pat, honestly. At
any rate, the fix is ugly as hell, so that the whole class will need to
be reviewed some time soon.
The basic idea is that there is now a mechanism to reset a JPEGImage to
the state it was in right after construction (no coefficients have been
read, and headers have just been taken in). This also involves turning
readCoefficients() into a no-op when coefficients have already been
read. (Otherwise big bad crashes take place. It is awful). The
implementation is currently about as clunky as it gets, so changes will
hopefully come.