A test for the blockiness calculation method.
We can't really verify that the formula is correctly implemented (other
than by looking at it), but we can ensure that the assumptions we make
about it are correct. Particularly, that an image containing an embedded
message, and that its cropped-to-image ratio are both larger of those of
its non-steganographic counterpart.