The permutations are now mostly stateless.
They no longer keep track of which particular index we're looking at, so
that it's possible to actually look back, and in general do not act as
iterators any more. It's still possible for them to be initialized
multiple times, thus being reseeded, which we want to get rid of soon.