Make the blockiness function actually measure blockiness...
As it turns out, I hadn't really thought anything through. The fact of
the matter is that the blockiness function measures, well, blockiness,
that is it measures inconsistencies between a data unit and its
neighbours. The way it was implemented was making all measurements
within the same data unit. It was basically a glorified off by one
error. Should be fixed now, although (hooray!!) there is NO WAY to
actually test this...