Re-write the blockiness algorithm.
It now actually does what it's supposed to. Previously, it attempted to
measure blockiness using the frequency domain. This is obviously bloody
impossible. It turns out that I'm an idiot and wasn't thinking straight.
At any rate, it should now work. More tests and refactoring for this are
forthcoming.
Additionally, make this blockiness thing operate exclusively with the
ROB, which is what we're interested in. This means we only have to
decompress the image once: we decompress it, calculate the blockiness,
crop it, decompress the crop, and get the blockiness for the crop.
Doing this really speeds us up.