Parallelize execution of the simulate() method.
Yay, this is good. We have to actually copy the JPEGImage for every
individual, which you'd think is kind of expensive but apparently
libjpeg is smart enough to figure out that we need a fast copy (from the
fact that no DCT coefficients have been accessed). The copy is necessary
because JPEGImages are about as far from thread safe as it gets, so we
want to try and encapsulate them into Individuals as much as humanly
possible.