Some other minor optimizations to the blockiness calculation.
Again, prevent a constant value from being re-computed over and over,
and remove the cropRow method so as to not grow the call stack. The code
is looking really messy, so hopefully this actually makes a difference.
Looking at some informal, ad-hoc, profiling it seems to help a tiny bit.