Split the EmbedProcedure into two different classes.
We now have one class for a real embed, one for a test embed; this way
we don't have to do that if(real) check for every single message bit.
There is some overhead associated  with doing this, but it's probably a
gain in the long run, although I've not seen any differences as of this
writing (I've yet to re-profile the app in any meaningful way).