Building up a gradle config that can actually compile the native side.
It actually works! Goes through the motions of swig and buidling the
shared library. Interface file had to be moved to simplify the building
a bit. Makefile was updated, since it still is the only way to place the
SWIG generated java files in the right places.
In the future, this same gradle project will create a JAR with the SWIG
java classes, so that we can pull it into the java side.