A gradle file for the Java project, and a root settings.gradle
Hooray! We can now issue a "gradle test" command from the root of the
project and have it actually compile our C++ code, our Java code, move
the libraries the right way and then test the Java code. This is very
good stuff. The stuff with the paths is very nasty currently, and ought
to be fixed soon. We also need some hooks into the clean methods just to
clean everything properly. And, we need to actually run the C++ unit
tests...