The coefficientAccessor's getLength() method is now unsigned.
This removes some sign warnings, and is overall a happier, more natural
declaration.
Unfortunately, this also means we have to declare some dicey arrays on
the Java side (unsigned ints in C++ become longs in Java). For the time
being this is just in the PM1Test, so it's okay because there we
actually have control over the image in use.